Feat : 新增前端畫面，thymeleaf搭配springsecurity

+++ src/main/java/com/rogerli/springmall/controller/ProductController.java
@@ -85,7 +85,10 @@ public class ProductController {
 
     @DeleteMapping("/products/{productId}")
     public ResponseEntity<?> deleteProduct(@PathVariable Integer productId){
-        productService.deleteProductById(productId);
+        Product product = productService.getProductById(productId);
+        if (product != null){
+            productService.deleteProductById(productId);
+        }
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 }
